Clean up the code

+++ Arduino-project/src/main.cpp
@@ -3,11 +3,9 @@
 
 // Components initialization
 Vehicle vehicle;
-long long lastTime = millis();
+unsigned long lastTime = millis();
 
-void setup()
-{
-  // put your setup code here, to run once:
+void setup() {
   pinMode(4, INPUT);
   Serial.begin(9600);
   Serial.println("Initializing vehicle");
@@ -16,11 +14,8 @@ void setup()
 }
 
 void loop() {
-  // put your main code here, to run repeatedly:
   //Serial.println("Updating vehicle");
   vehicle.update();
-  //pinMode(4, INPUT);
-  //Serial.println(analogRead(4));
   unsigned long currentTime = millis();
   if (currentTime - lastTime < 10) delay(10 - (currentTime - lastTime));
   currentTime = millis();
